Move language host functionality to module.
