+++ src/service-host/interfaces.ts
